more logging and stuff
