+++ baldrick/plugins/github_milestones.py
@@ -1,3 +1,5 @@
+from loguru import logger
+
 from baldrick.plugins.github_pull_requests import pull_request_handler
 
 MISSING_MESSAGE = 'This pull request has no milestone set.'
@@ -9,10 +11,12 @@ def process_milestone(pr_handler, repo_handler):
     """
     A very simple set a failing status if the milestone is not set.
     """
-
     mc_config = pr_handler.get_config_value("milestones", {})
     if not mc_config.get('enabled', False):
+        logger.debug("Skipping milestone plugin as disabled in config")
         return
+
+    logger.debug(f"Checking milestones on {pr_handler.repo}#{pr_handler.number}")
 
     fail_message = mc_config.get("missing_message", MISSING_MESSAGE)
     pass_message = mc_config.get("present_message", PRESENT_MESSAGE)
